refactor: move closeTabsLeft and closeTabsRight functions into separate files
