+++ packages/renderer-worker/src/parts/ViewletMain/ViewletMain.js
@@ -618,55 +618,6 @@ export const handleFocusChange = (state, isFocused) => {
   return state
 }
 
-export const closeOthers = async (state) => {
-  if (state.focusedIndex === state.activeIndex) {
-    // view is kept the same, only tabs are closed
-    await RendererProcess.invoke(
-      /* Viewlet.send */ 'Viewlet.send',
-      /* id */ state.uid,
-      /* method */ 'closeOthers',
-      /* keepIndex */ state.focusedIndex,
-      /* focusIndex */ state.focusedIndex
-    )
-  } else {
-    // view needs to be switched to focused index
-    await RendererProcess.invoke(
-      /* Viewlet.send */ 'Viewlet.send',
-      /* id */ state.uid,
-      /* method */ 'closeOthers',
-      /* keepIndex */ state.focusedIndex,
-      /* focusIndex */ state.focusedIndex
-    )
-  }
-  state.editors = [state.editors[state.focusedIndex]]
-  state.activeIndex = 0
-  state.focusedIndex = 0
-}
-
-export const closeTabsRight = async (state) => {
-  if (state.focusedIndex >= state.activeIndex) {
-    // view is kept the same, only tabs are closed
-    await RendererProcess.invoke(/* Viewlet.send */ 'Viewlet.send', /* id */ state.uid, /* method */ 'closeTabsRight', /* index */ state.focusedIndex)
-  } else {
-    // view needs to be switched to focused index
-    await RendererProcess.invoke(/* Viewlet.send */ 'Viewlet.send', /* id */ state.uid, /* method */ 'closeTabsRight', /* index */ state.focusedIndex)
-  }
-  state.editors = state.editors.slice(0, state.focusedIndex + 1)
-  state.activeIndex = state.focusedIndex
-}
-
-export const closeTabsLeft = async (state) => {
-  if (state.focusedIndex <= state.activeIndex) {
-    // view is kept the same, only tabs are closed
-    await RendererProcess.invoke(/* Viewlet.send */ 'Viewlet.send', /* id */ state.uid, /* method */ 'closeTabsLeft', /* index */ state.focusedIndex)
-  } else {
-    // view needs to be switched to focused index
-    await RendererProcess.invoke(/* Viewlet.send */ 'Viewlet.send', /* id */ state.uid, /* method */ 'closeTabsLeft', /* index */ state.focusedIndex)
-  }
-  state.editors = state.editors.slice(state.focusedIndex)
-  state.activeIndex = state.focusedIndex
-}
-
 export const resize = (state, dimensions) => {
   const { editors } = state
   const x = dimensions.x
